Transcriber-bot/requirements.txt
openai-whisper
pyaudio
sounddevice
keyboard
numpy
scipy
cl